security-threat-customs-reports/script-deprecated-report-updater/src/requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
greenlet==3.0.3
idna==3.7
packaging==24.1
playwright==1.44.0
pyee==11.1.0
python-dotenv==1.0.1
requests==2.32.3
typing_extensions==4.12.2
urllib3==2.2.2
